2008-01-07  Jorn Baayen  <jorn@openedhand.com>
	* libgupnp/gvalue-util.c: (gvalue_util_set_value_from_string):

	Use atol() for 64-bit int parsing.



git-svn-id: https://svn.o-hand.com/repos/gupnp/trunk/gupnp@666 d8cb91d7-bff9-0310-92b9-80b65e4482b2

+++ libgupnp/gvalue-util.c
@@ -66,13 +66,13 @@ gvalue_util_set_value_from_string (GValue     *value,
                 break;
 
         case G_TYPE_INT64:
-                i = atoi (str);
+                i = atol (str);
                 g_value_set_int64 (value, (gint64) i);
 
                 break;
 
         case G_TYPE_UINT64:
-                i = atoi (str);
+                i = atol (str);
                 g_value_set_uint64 (value, (guint64) i);
 
                 break;
